+++ maketoken.js
@@ -1,14 +1,4 @@
-const { clientOptions, connectionOptions } = require('./options');
-const { encrypt, deepMerge, loadConfig } = require('./utils');
+const { makeToken } = require('./lib/token');
+const { loadConfig } = require('./lib/utils');
 
-const credentials = loadConfig(__dirname);
-const options = deepMerge(connectionOptions, {
-  connection: {
-    settings: {
-      username: 'abc',
-      password: credentials.password,
-    },
-  },
-});
-
-console.log(encrypt(options, clientOptions.crypt.cypher, clientOptions.crypt.key));
+console.log(makeToken(loadConfig(__dirname)));
