+++ package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "test": "jest --coverage --detectOpenHandles",
-    "lint": "eslint *.js"
+    "lint": "eslint *.js lib/*.js"
   },
   "repository": {
     "type": "git",
